Create montecarloscc_prtp_eta.jl
This function is derived from compute_sc() from new_marginaldamages.jl it is almost the same, but it allows eta and prtp to be uncertain, the distribution function assigned is a uniform that goes from the value assigned by default more or less 10%, following the ASG methodology used in Anderson et al (2012). The method sc_from_mm is copied from new_marginaldamages.jl the function _compute_sc_from_mm().